rename use-new-activity
